Adds os-clr-on-clr-dev and dependent -dev bundles.
Last phased-in load of automated -dev bundles for a while.  This
finishes creating -dev bundles for all include()'s in os-clr-on-clr,
which covers most of the operating system.  There's a reasonable
amount of new stuff here but the vast majority is still coming from
include()'s that are already present such as os-core-dev.