Deprecate and 'delete' database-mariadb
The database-basic bundle is intended to be used for running a database
instead of an application specific bundle.

Currently openstack-all-in-one (itself used by bat) is the only bundle
using this bundle. Changing that bundle to use the new database-basic
bundle is an option but holding off on that for now.