Deprecate os-utils
os-utils ended up as a very hard to quantify grab bag of packages and
includes. Instead of of this, bundles should largely be installed to get
specific functionality. To that end specific packages largely have been
moved to sysadmin-basic as they fit in with that use otherwise
functionality in this bundle is available already in seperate bundles
that have a more obvious purpose already.