Rework os-clr-on-clr
The os-clr-on-clr bundle is a one stop shop bundle for all Clear Linux
development and has been reworked to better include bundles instead of
adding one or two different components of a bundle directly.

This change also makes more use of the doc bundles and wants a new
storage-utils-doc bundle created before another commented out change is
added.

Finally there is a little rework of other included bundles to better
capture items that should have already been in them but were in
os-clr-on-clr instead.