Split webapp packages into their own bundles
This is the start of removing the lamp-basic and lnmp-basic bundles
which are content defined rather than functional. Instead users can
install database, scripting language and web-server bundles individually
to provide similar feature sets.

In the future there may also be a bundle including all the pieces to
make a pure include webapp bundle as well.