Deprecate the lamp-basic and lnmp-basic bundles
The lamp-basic and lnmp-basic bundles are overly content
defined. Currently the major use case internally for lamp-basic is
os-testsuite-phoronix which is changed to include the database-basic,
php-basic and web-server-basic bundles and the bndl-lamp-basic
package (for the enable symlinks). In general enable symlinks for
services that would conflict should not be in a generic bundle, only
specific use bundles (like os-testsuite-phoronix).