Add console user focused meta bundle for usability
There has been multiple requests to have a single bundle that covers a
number of different workflows that a single system administrator user
will want to run.

The user-basic bundle is an attempt to provide these features by
including multiple different functional groups for ease of documentation
and usability.

As a note a meta bundle it is defined by its includes only and does not
itself include packages directly.