move connect-proxy helper from os-clr-on-clr to network-basic
The connect-proxy is a simple, compiled, c-based helper for proxy
traversal.  It can be used as the ProxyCommand in a .ssh/config, for
example enabling proxy traversal for an ssh based git connection to a
git repo server.

This is a small, commonly useful utility and should not be limited to
os-clr-on-clr.  The network-basic bundle is a more natural location and
that is included by os-clr-on-clr.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>