Use base package instead of specific subpackages
For cryptograhy we aren't particuarly required to be extremely size
sensative so adding the base package for openssh (giving users docs and
libraries for instance) seems to be the right move. Also sort the
package list.

Note nss-bin was not converted to nss as nss isn't itself a valid
package and nss-bin is instead the main installable package target.