Support building cgo programs with go-basic bundle
On a clean install of Clear Linux, just installing the go-basic bundle
does not allow cgo programs to be compiled.  Because this use case is
something go supports as a part of the language, but because the
functionality for compiling C programs is best encapsulated in the
c-basic bundle, this patch modifies go-basic to include c-baisc.