Revert "Support building cgo programs with go-basic bundle"
Language bundles are designed to be pluggable/modular to put the user in
control of exactly what they need for their task at hand.  They do not
try to predict the needs of the user and their programs because it
varies, case-by-case.  If users need to build cgo programs, then they
should install both the go-basic and c-basic bundles.