hpc-utils: new bundle, rdma/IB tool rework
We're dropping the old librdmacm for rdma-core, which includes
a bunch of libraries that are now obsolete. The rebuilding of
packages is already in progress, but the bundles need unbreaking
as well as having a new bundle to collect these tools into.

This bundle is WIP. Its content may move to one of the other
bundles in the future.