Add `clr-service-restart` binary-only to os-core-update
This ~16k binary will, at a later point in time, be used to
automatically restart services that need that. With this binary on
each system, we'll be able to figure out quickly what services we
should add to the `whitelist` so that we can make a reasonably safe
list of daemons that are safe to have restart automatically.