Remove jedi-vim and python-basic from editors
jedi-vim does not work well at all in Clear Linux. Even when it is
working well, it slows vim down considerably. Even if you could get
around all that, it pulls in all of python-basic just to use an arguably
not very useful broken plugin. If you want this yourself, just use
pathogen or similar to install it locally.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>