Revert "Update the qt-basic bundle"
This reverts commit 8d947b550f084a7f85e5a5b836550e646d3bbc25.

I asked this commit get pushed prior to review to avoid having to
untag the qt packages since they have conflicts with the
qt-everywhere-opensource-src package. However, after the fact it
seems that a lot of other packages have the qt-everywhere package
as a dependency, and it cannot be blocked yet. So this commit is
being reverted so that it can go through the typical review
process.