Remove unused packages from mixer bundle
With the deprecation of the old implementations of mixer functionality a
lot of dependencies can be removed. Now that Clear Linux uses DNF, the
yum dependencies can also be removed. We have to keep
bundle-chroot-builder around a little bit longer until the mixer package
is fixed to provide the builder.conf in the right place on the system.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>