Move the git pundle to full bundle
Due to git being included in a number of other bundles and git itself
pulling in multiple package as dependencies that are already in their
own bundles, having git be able to include other bundles is a big win
for swupd performance. So move git from a pundle to bundle and add the
corresponding includes.