Add stress-ng to performance tools
stress-ng will stress test a computer system in various selectable ways. It
was designed to exercise various physical subsystems of a computer as well as
the various operating system kernel interfaces.

stress-ng is not recommended to be on os-testsuite-phoronix since the
idea of that bundle is to have everything for phoronix tests suites.
This is not a phoronix test and some power and performance developers
might find necessary to have a tool to stress the system with a tool like
stress-ngwith :

  * over 190 stress tests
  * over 50 CPU specific stress tests that exercise floating point, integer,
    bit manipulation and control flow
  * over 20 virtual memory stress tests

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>