Add an x11-tools bundle collecting X11-only tools
New tools in this commit:
 - setxkbmap
 - setxkbmap
 - xdpyinfo
 - xinput
 - xmodmap
 - xprop
 - xwininfo

When you develop graphical applications, getting information from
windows (xprop, xwininfo) is generally useful. You'd use xdpyinfo to
find out if your X session is working and it would have been useful for
me last night. setxkbmap with the -print option is good to determine why
some extended keys may not be working.

xinput comes in handy for me because when I turn RFKill on for WiFi,
the touchpad loses multi-finger and tapping support until I toggled it
off and on.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>