Add `x11-server` bundle, and include it where appropriate.
This is a first pass at adding a generic `x11-server` bundle. It
contains just the bare Xorg/wayland components, but no DM or toolkit
support. This should be a building block for desktop environments to
provide a basic, functional /usr/bin/X that can work through either
gdm, another login manager, or just `startx` where appropriate.

This isn't perfect just yet. There are many libraries like libva that
may or may not end up in the new bundle, but perhaps they are better
off in a media support bundle.