Clear Installer support bundles
These bundles are optionally added to a target install image by the
Clear Linux OS Installer based on configuration specifications.

If a system locale other than the default of 'en_US.UTF-8' is specified,
the 'locales' bundle will be installed on the target. ~204MB

If a keyboard other than the default of 'us' is specified, the
'kbd' package bundle will be installed on the target. ~4.7MB

If a time zone other than the default of 'UTC' is specified,
the 'tzdata' package bundle will be installed on the target. ~1MB

All three of these bundles will be added to the Clear Linux OS
Installer image used for interactive installations, but only
the required bundles are installed on the target to reduce image size.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>