bundles/games: Organise all SDL libraries into sections
We now organise based on SDL1 or SDL2 ABI/API, and further organise by
native and 32-bit builds. This helps ensure we keep all variants of SDL
present for games/Steam compatibility as the previous list was a bit of
a mess. Additionally we didn't include native 64-bit versions before.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>