bundles/games: Add compat-curl-gnutls-soname4
This is required for binary compatibility with a large number of games
including Don't Starve, as they have incorrectly assumed the Debian-specific
soname is standard, and fail to bundle it with the games.

We rectify this with a compatibility package that is built against gnutls
and the soname modified to reflect the Debian convention of modified
soname.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>