Add a kdiff3 as a pundle and as part of desktop-kde
kdiff3 is a very useful even outside of Plasma desktop, as it's a
nice, powerful and intuitive driver for "git mergetool". That's how
I've been dealing with Git conflicts for the past 12 years.

This is the unfinished Qt5/KF5 port of kdiff3, since Clear does not
have Qt4.

(The 3 indicates 3-way diff, not KDE3)

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>