Split minetest into client and server.
This should allow headless servers to run without any X11, as
is needed. The server pulls in the core game code, which is needed
for most basic servers. The client then includes the server bundle
and should thus be complete as well. I've made the games bundle
include the client, completing this change.