Update README and bundle-header template
Update bundle-header to include Pending-Delete and Deprecated as two
status types that impact how bundle deletion is handled during a
format bump.

It is intended that bundles that are being replaced must go
through a Pending-Delete phase where the replacement bundle is
included and a format bump occurs that will cause the include to
become installed. From there the bundle can move to Pending-Delete
and be emptied (containing only the placeholder filesystem package).