os-core: Prepare for bundle re-factorization
This change is a precursor to the removal of content from os-core that
specifically was identified as being surplus to the requirements of a
container providing services.

The content to be removed has been placed in the os-core-plus bundle
which enables running systemd as PID1 and a number of other very basic
programs and capabilities.

os-core-update has been modified to include os-core-plus as the Clear
Linux images and installs (outside of some specific containers) all
have os-core-update by default.

This is intended to be a change which has no impact on the operation
of deployed systems and just enables future work to make smaller
container images.