os-core: Remove content provided by os-core-plus
Make the removal described in 37edd79c3870b2cc9220865df5b35484059b3929
to reduce os-core size specifically with a container use case in
mind. This change should be a no-op for users with os-core-update
installed which is any non-container image currently being
produced. If you making images not using os-core-update but still
intending to use systemd, ensure os-core-plus is added.