machine-learning-basic: Drop aeon as it is EOL
from aeon GitHub repo [1]

DISCONTINUATION OF PROJECT. This project will no longer be maintained
by Intel. Intel will not provide or guarantee development of or support
for this project, including but not limited to, maintenance, bug fixes,
new releases or updates. Patches to this project are no longer accepted
by Intel. If you have an ongoing need to use this project, are interested
in independently developing it, or would like to maintain patches for the
community, please create your own fork of the project.

[1] https://github.com/NervanaSystems/aeon