Move to manual maintenance for `devpkg-gnu-efi`
The `devpkg-gnu-efi` bundle is created automatically via the `auto.devpkgs`
mechanism. However, that mechanism does not add the `-staticdev` subpackage to
the bundle. This is problematic as it contains the required (static) libraries.
`gnu-efi` does not provide shared objects.

Because `gnu-efi` is rather unique in this, we drop the auto.devpkgs mechanism,
maintain the devpkg-gnu-efi bundle definition manually and add the
`gnu-efi-staticdev` subpackage to it.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>