[mixer]: Remove containers-basic include
According to the release notes for mixer 6.0.0:

"Built in support for Docker has also been removed. The `--native` flag
has been deprecated and all commands now run in a native setup. Docker
support was introduced to allow auto-format-bump feature to be
implemented, but current mixer implementation does not require it
anymore to work. External docker images, like mixer-ci, should be used
as a replacement in use cases where the built-in docker support was used
to create mix in incompatible host OSes."

Signed-off-by: George T Kramer <george.t.kramer@intel.com>